fix 설문 제작(질문 1개 이하일때 , 보기 1개이하일때 삭제 불가하도록

+++ scratch/src/pages/SurveyMake.js
@@ -14,7 +14,7 @@ import QuestionMakeList from '../components/Survey/QuestionMakeList';
 import QuestionResultList, { MAKE } from '../components/Survey/QuestionResultList';
 import produce from 'immer';
 
-import {Menu,MenuItem,Button,IconButton} from '@mui/material'
+import {Menu,MenuItem,Button,IconButton,Tooltip} from '@mui/material'
 import DeleteIcon from '@mui/icons-material/Delete';
 import { TO_BACKEND_OPTION } from '../components/redux/Slices/SurveyOptionSlice';
 
@@ -130,14 +130,30 @@ export function PlusButton({id}){
 }
 
 export function DeleteButton({id}){
+    const question = useSelector((state)=>state.surveyMake.question);
     const dispatch = useDispatch();
-    return(
-      <div>
-        <IconButton variant="contained"  size="small" 
-        onClick={() => {dispatch(DELETE({id:id}));
-        }}><DeleteIcon /></IconButton>
-       </div>
-    );
+
+    if( question.length >= 2){
+      return(
+        <div>
+          <IconButton variant="contained"  size="small" 
+          onClick={() => {dispatch(DELETE({id:id}));
+          }}><DeleteIcon/></IconButton>
+         </div>
+      );
+
+    }
+    else{
+      return(
+        <div>
+          <Tooltip title="질문은 하나 이상 필요합니다!">
+          <IconButton variant="contained"  size="small" ><DeleteIcon/></IconButton>
+            </Tooltip>
+         </div>
+      );
+
+    }
+    
 }
